+++ requirements.txt
@@ -1,4 +1,4 @@
 igraph==0.11.4
 graphviz==0.20.1
 cairocffi==1.7.1  # for igraph.plot
-rough-theory==0.0.1
+rough-theory==0.0.2
