+++ src/regime/nodes/meta.py
@@ -3,6 +3,7 @@ Defines the metaclass for Regime Nodes. This metaclass automatically collects hy
 the class (assuming it has been decorated with the hyperparameter decorator) and stores them in a
 dictionary.
 """
+
 from abc import ABC
 
 
@@ -12,6 +13,10 @@ class HyperparameterMeta(type(ABC)):  # was originally just: type
     has been decorated with the hyperparameter decorator) and stores them in a dictionary.
     """
 
+    _hyperparameters = {}  # only needed to avoid a pylint error (no-member)
+
+    # C0204: Metaclass class method __new__ should have 'mcs' as first argument
+    # (bad-mcs-classmethod-argument)
     def __new__(mcs, name, bases, dct):
         dct["_hyperparameters"] = {}
         for _, attr in dct.items():  # attr_name, attr
